Various fixes
Chip6502.cpp
* Additional_Cycle1 & 2 were reversed
* Missing SetFlag inside Clock

Chip6502.h
* pc was of wrong size. needed to be uint16_t
* setFlag -> SetFlag

Bus.h
* MAX_MEM was wrong size, now uint32_t

+++ headers/Bus.h
@@ -4,7 +4,7 @@
 #include <cstdint>
 #include "Chip6502.h"
 
-const uint8_t MAX_MEM = 1024 * 64;
+const uint32_t MAX_MEM = 1024 * 64;
 
 class Bus 
 {
@@ -21,6 +21,7 @@ public:
     uint8_t read(uint16_t addr, bool bReadOnly = false);
 
 private:
+
     // Clean Ram
     void clear();
 };
